Cleaned and updated version 31 oct. 2013
I'll try to give a complete description of what I did since the last
commit:
- I added the "Preface" part that Moti wrote (translated of course)
- I corrected the text thanks to the corrections of Mariza, there should
not be much spelling mistakes anymore
- I went through all the images to check if the title was clear enough
and in English. I also looked a bit which one I modified from Moti's
version or added. Most of them I just took from what Moti did, then I
added a few new ones to match the modifications I did in the text. I
also modified some of them to be more suitable in the text (minor text
modifications or simply esthetic considerations) The only ones I really
modified which could be put in Moti's version is the "action-colors-up"
and "action-colors-down" that I changed from black background to white
one (in order to see the wheels on the bottom ones) The new files are
called "action-colors-up-white" and "action-colors-down-white".
- I finally went through the image folder to find if some of them were
not used. The folder I have now is the strict minimum for Moti's version
and mine. Both compile without problems.

I hope it's ok this way, I will stay until the end of November (probably
more, but not sure yet) Anyway, I can be reached by mail anytime to make
it clean ;o)